+++ packages/safe-apps-provider/src/provider.ts
@@ -1,7 +1,7 @@
-import SafeAppsSDK, { SafeInfo, Web3TransactionObject } from '@gnosis.pm/safe-apps-sdk';
+import SafeAppsSDK, { SafeInfo, TransactionStatus, Web3TransactionObject } from '@safe-global/safe-apps-sdk';
 import { EventEmitter } from 'events';
-import { EIP1193Provider } from './types';
-import { getLowerCase } from './utils';
+import { EIP1193Provider, GetCallsParams, GetCallsResult, SendCallsParams, SendCallsResult } from './types';
+import { getLowerCase, numberToHex } from './utils';
 
 // The API is based on Ethereum JavaScript API Provider Standard. Link: https://eips.ethereum.org/EIPS/eip-1193
 export class SafeAppProvider extends EventEmitter implements EIP1193Provider {
@@ -38,7 +38,7 @@ export class SafeAppProvider extends EventEmitter implements EIP1193Provider {
 
       case 'net_version':
       case 'eth_chainId':
-        return `0x${this.chainId.toString(16)}`;
+        return numberToHex(this.chainId);
 
       case 'personal_sign': {
         const [message, address] = params;
@@ -47,9 +47,10 @@ export class SafeAppProvider extends EventEmitter implements EIP1193Provider {
           throw new Error('The address or message hash is invalid');
         }
 
-        await this.sdk.txs.signMessage(message);
+        const response = await this.sdk.txs.signMessage(message);
+        const signature = 'signature' in response ? response.signature : undefined;
 
-        return '0x';
+        return signature || '0x';
       }
 
       case 'eth_sign': {
@@ -59,9 +60,10 @@ export class SafeAppProvider extends EventEmitter implements EIP1193Provider {
           throw new Error('The address or message hash is invalid');
         }
 
-        await this.sdk.txs.signMessage(messageHash);
+        const response = await this.sdk.txs.signMessage(messageHash);
+        const signature = 'signature' in response ? response.signature : undefined;
 
-        return '0x';
+        return signature || '0x';
       }
 
       case 'eth_signTypedData':
@@ -73,21 +75,31 @@ export class SafeAppProvider extends EventEmitter implements EIP1193Provider {
           throw new Error('The address is invalid');
         }
 
-        await this.sdk.txs.signTypedMessage(parsedTypedData);
-
-        return '0x';
+        const response = await this.sdk.txs.signTypedMessage(parsedTypedData);
+        const signature = 'signature' in response ? response.signature : undefined;
+        return signature || '0x';
       }
 
       case 'eth_sendTransaction':
+        // `value` or `data` can be explicitly set as `undefined` for example in Viem. The spread will overwrite the fallback value.
         const tx = {
-          value: '0',
-          data: '0x',
           ...params[0],
+          value: params[0].value || '0',
+          data: params[0].data || '0x',
         };
+
+        // Some ethereum libraries might pass the gas as a hex-encoded string
+        // We need to convert it to a number because the SDK expects a number and our backend only supports
+        // Decimal numbers
+        if (typeof tx.gas === 'string' && tx.gas.startsWith('0x')) {
+          tx.gas = parseInt(tx.gas, 16);
+        }
 
         const resp = await this.sdk.txs.send({
           txs: [tx],
+          params: { safeTxGas: tx.gas },
         });
+
         // Store fake transaction
         this.submittedTxs.set(resp.safeTxHash, {
           from: this.safe.safeAddress,
@@ -180,8 +192,108 @@ export class SafeAppProvider extends EventEmitter implements EIP1193Provider {
       case 'wallet_requestPermissions':
         return this.sdk.wallet.requestPermissions(params[0]);
 
-      case 'safe_useOffChainSignatures': {
-        return this.sdk.safe.useOffChainSignatures(params[0]);
+      case 'safe_setSettings':
+        return this.sdk.eth.setSafeSettings([params[0]]);
+
+      case 'wallet_sendCalls': {
+        const { from, calls, chainId }: SendCallsParams = params[0];
+
+        if (chainId !== numberToHex(this.chainId)) {
+          throw new Error(`Safe is not on chain ${chainId}`);
+        }
+
+        if (from !== this.safe.safeAddress) {
+          throw Error('Invalid from address');
+        }
+
+        const txs = calls.map((call, i) => {
+          if (!call.to) {
+            throw new Error(`Invalid call #${i}: missing "to" field`);
+          }
+          return {
+            to: call.to,
+            data: call.data ?? '0x',
+            value: call.value ?? numberToHex(0),
+          };
+        });
+
+        const { safeTxHash } = await this.sdk.txs.send({ txs });
+
+        const result: SendCallsResult = {
+          id: safeTxHash,
+        };
+
+        return result;
+      }
+
+      case 'wallet_getCallsStatus': {
+        const safeTxHash: GetCallsParams = params[0];
+
+        const CallStatus: {
+          [key in TransactionStatus]: GetCallsResult['status'];
+        } = {
+          [TransactionStatus.AWAITING_CONFIRMATIONS]: 100,
+          [TransactionStatus.AWAITING_EXECUTION]: 100,
+          [TransactionStatus.SUCCESS]: 200,
+          [TransactionStatus.CANCELLED]: 400,
+          [TransactionStatus.FAILED]: 500,
+        };
+
+        const tx = await this.sdk.txs.getBySafeTxHash(safeTxHash);
+
+        const result: GetCallsResult = {
+          version: '1.0',
+          id: safeTxHash,
+          chainId: numberToHex(this.chainId),
+          status: CallStatus[tx.txStatus],
+        };
+
+        // Transaction is queued
+        if (!tx.txHash) {
+          return result;
+        }
+
+        // If transaction is executing, receipt is null
+        const receipt = await this.sdk.eth.getTransactionReceipt([tx.txHash]);
+        if (!receipt) {
+          return result;
+        }
+
+        const calls =
+          tx.txData?.dataDecoded?.method !== 'multiSend'
+            ? 1
+            : // Number of batched transactions
+              tx.txData.dataDecoded.parameters?.[0].valueDecoded?.length ?? 1;
+
+        // Typed as number; is hex
+        const blockNumber = Number(receipt.blockNumber);
+        const gasUsed = Number(receipt.gasUsed);
+
+        result.receipts = Array(calls).fill({
+          logs: receipt.logs,
+          status: numberToHex(tx.txStatus === TransactionStatus.SUCCESS ? 1 : 0),
+          blockHash: receipt.blockHash,
+          blockNumber: numberToHex(blockNumber),
+          gasUsed: numberToHex(gasUsed),
+          transactionHash: tx.txHash,
+        });
+
+        return result;
+      }
+
+      case 'wallet_showCallsStatus': {
+        // Cannot open transaction details page via SDK
+        throw new Error(`"${request.method}" not supported`);
+      }
+
+      case 'wallet_getCapabilities': {
+        return {
+          [numberToHex(this.chainId)]: {
+            atomicBatch: {
+              supported: true,
+            },
+          },
+        };
       }
 
       default: