+++ packages/safe-apps-sdk/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  clearMocks: true,
-  moduleDirectories: ['node_modules', 'src'],
-  roots: ['src'],
-  verbose: true,
-  setupFiles: ['./jest/setup.ts'],
-  testEnvironment: 'jsdom',
-};