+++ packages/safe-apps-sdk/src/safe/safe.test.ts
@@ -1,11 +1,25 @@
-import SDK from '../sdk';
-import { SafeInfo } from '../types';
-import { Methods } from '../communication/methods';
-import { PostMessageOptions } from '../types';
-import { PermissionsError } from '../types/permissions';
-import { Wallet } from '../wallet';
+import SDK from '../sdk.js';
+import { SafeInfoExtended } from '../types/index.js';
+import { Methods } from '../communication/methods.js';
+import { PostMessageOptions } from '../types/index.js';
+import { PermissionsError } from '../types/permissions.js';
+import { Wallet } from '../wallet/index.js';
 
 const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
+
+export const ExtendedSafeInfo: SafeInfoExtended = {
+  chainId: 4,
+  safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
+  nonce: 1,
+  implementation: '0x3E5c63644E683549055b9Be8653de26E0B4CD36E',
+  modules: null,
+  fallbackHandler: null,
+  version: '1.3.0',
+  guard: null,
+  owners: [],
+  threshold: 1,
+  isReadOnly: false,
+};
 
 describe('Safe Apps SDK safe methods', () => {
   const sdkInstance = new SDK();
@@ -33,16 +47,7 @@ describe('Safe Apps SDK safe methods', () => {
   describe('SDK.safe.calculateMessageHash', () => {
     test('Should generate correct EIP-191 message hash', () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       // to test message/hash I signed a test message on rinkeby
       // https://dashboard.tenderly.co/tx/rinkeby/0x9308fb61d9f4282080334e3f35b357fc689e06808b8ad2817536813948e3720d
       const message = 'approve rugpull';
@@ -53,20 +58,53 @@ describe('Safe Apps SDK safe methods', () => {
     });
   });
 
+  describe('SDK.safe.calculateTypedMessageHash', () => {
+    test('Should generate correct EIP-712 message hash', () => {
+      const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
+
+      const typedMessage = {
+        domain: {
+          name: 'Ether Mail',
+          version: '1',
+          chainId: 1,
+          verifyingContract: '0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC',
+        },
+        types: {
+          Person: [
+            { name: 'name', type: 'string' },
+            { name: 'wallet', type: 'address' },
+          ],
+          Mail: [
+            { name: 'from', type: 'Person' },
+            { name: 'to', type: 'Person' },
+            { name: 'contents', type: 'string' },
+          ],
+        },
+        message: {
+          from: {
+            name: 'Cow',
+            wallet: '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
+          },
+          to: {
+            name: 'Bob',
+            wallet: '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
+          },
+          contents: 'Hello, Bob!',
+        },
+      };
+      const expectedHash = '0xbe609aee343fb3c4b28e1df9e632fca64fcfaede20f02e86244efddf30957bd2';
+      const hash = sdkInstance.safe.calculateTypedMessageHash(typedMessage);
+
+      expect(hash).toEqual(expectedHash);
+    });
+  });
+
   describe('SDK.safe.check1271Signature', () => {
     test('Should send a valid message to the interface', async () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
-      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // ethers.utils.formatBytes32String('approve rugpull')
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
+      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // stringToHex('approve rugpull', { size: 32 })
       // @ts-expect-error method is private but we are testing it
       sdkInstance.safe.check1271Signature(message);
       await sleep(200);
@@ -92,16 +130,7 @@ describe('Safe Apps SDK safe methods', () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
       // @ts-expect-error method is private but we are testing it
       const rpcCallSpy = jest.spyOn(sdkInstance.safe.communicator, 'send');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       rpcCallSpy.mockImplementationOnce(() =>
         Promise.resolve({
           id: '1',
@@ -110,7 +139,7 @@ describe('Safe Apps SDK safe methods', () => {
         }),
       );
 
-      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // ethers.utils.formatBytes32String('approve rugpull')
+      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // stringToHex('approve rugpull', { size: 32 })
       // @ts-expect-error method is private but we are testing it
       expect(await sdkInstance.safe.check1271Signature(message)).toEqual(true);
     });
@@ -119,19 +148,10 @@ describe('Safe Apps SDK safe methods', () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
       // @ts-expect-error method is private but we are testing it
       const rpcCallSpy = jest.spyOn(sdkInstance.safe.communicator, 'send');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       rpcCallSpy.mockImplementationOnce(() => Promise.reject(new Error('Hash not approved')));
 
-      const message = '0x68616c6c6f000000000000000000000000000000000000000000000000000000'; // ethers.utils.formatBytes32String('hallo')
+      const message = '0x68616c6c6f000000000000000000000000000000000000000000000000000000'; // stringToHex('hallo')
       // @ts-expect-error method is private but we are testing it
       expect(await sdkInstance.safe.check1271Signature(message)).toEqual(false);
     });
@@ -140,17 +160,8 @@ describe('Safe Apps SDK safe methods', () => {
   describe('SDK.safe.check1271SignatureBytes', () => {
     test('Should send a valid message to the interface', async () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
-      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // ethers.utils.formatBytes32String('approve rugpull')
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
+      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // stringToHex('approve rugpull', { size: 32 })
       // @ts-expect-error method is private but we are testing it
       sdkInstance.safe.check1271SignatureBytes(message);
       await sleep(200);
@@ -176,16 +187,7 @@ describe('Safe Apps SDK safe methods', () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
       // @ts-expect-error method is private but we are testing it
       const rpcCallSpy = jest.spyOn(sdkInstance.safe.communicator, 'send');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       rpcCallSpy.mockImplementationOnce(() =>
         Promise.resolve({
           id: '1',
@@ -194,7 +196,7 @@ describe('Safe Apps SDK safe methods', () => {
         }),
       );
 
-      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // ethers.utils.formatBytes32String('approve rugpull')
+      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // stringToHex('approve rugpull', { size: 32 })
       // @ts-expect-error method is private but we are testing it
       expect(await sdkInstance.safe.check1271SignatureBytes(message)).toEqual(true);
     });
@@ -203,19 +205,10 @@ describe('Safe Apps SDK safe methods', () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
       // @ts-expect-error method is private but we are testing it
       const rpcCallSpy = jest.spyOn(sdkInstance.safe.communicator, 'send');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       rpcCallSpy.mockImplementationOnce(() => Promise.reject(new Error('Hash not approved')));
 
-      const message = '0x68616c6c6f000000000000000000000000000000000000000000000000000000'; // ethers.utils.formatBytes32String('hallo')
+      const message = '0x68616c6c6f000000000000000000000000000000000000000000000000000000'; // stringToHex('hallo')
       // @ts-expect-error method is private but we are testing it
       expect(await sdkInstance.safe.check1271SignatureBytes(message)).toEqual(false);
     });
@@ -225,7 +218,7 @@ describe('Safe Apps SDK safe methods', () => {
     test('Should call SDK.safe.isMessageHashSigned with a hash of the message', () => {
       const isMessageHashSignedSpy = jest.spyOn(sdkInstance.safe, 'isMessageHashSigned');
 
-      // ethers.utils.formatBytes32String('approve rugpull')
+      // stringToHex('approve rugpull', { size: 32 })
       const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000';
       const expectedHash = '0xaae9257b8ff1c926ac3cdf36923661de4e81bf934e38958beeede3519aa18b08';
 
@@ -237,18 +230,9 @@ describe('Safe Apps SDK safe methods', () => {
   describe('SDK.safe.isMessageHashSigned', () => {
     test('Should send call messages to check the message the interface', async () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
-      safeInfoSpy.mockImplementation(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
 
-      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // ethers.utils.formatBytes32String('approve rugpull')
+      const message = '0x617070726f76652072756770756c6c0000000000000000000000000000000000'; // stringToHex('approve rugpull', { size: 32 })
 
       sdkInstance.safe.isMessageHashSigned(message);
       await sleep(200);
@@ -275,20 +259,11 @@ describe('Safe Apps SDK safe methods', () => {
       const safeInfoSpy = jest.spyOn(sdkInstance.safe, 'getInfo');
       // @ts-expect-error method is private but we are testing it
       const check1271SignatureSpy = jest.spyOn(sdkInstance.safe, 'check1271Signature');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       // @ts-expect-error ts fails to infer the return type because of a private method
       check1271SignatureSpy.mockImplementationOnce(() => Promise.resolve(true));
 
-      // ethers.utils.formatBytes32String('approve rugpull')
+      // stringToHex('approve rugpull', { size: 32 })
       const message = sdkInstance.safe.calculateMessageHash(
         '0x617070726f76652072756770756c6c0000000000000000000000000000000000',
       );
@@ -303,22 +278,13 @@ describe('Safe Apps SDK safe methods', () => {
       const check1271SignatureSpy = jest.spyOn(sdkInstance.safe, 'check1271Signature');
       // @ts-expect-error method is private but we are testing it
       const check1271SignatureBytesSpy = jest.spyOn(sdkInstance.safe, 'check1271SignatureBytes');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       // @ts-expect-error ts fails to infer the return type because of a private method
       check1271SignatureSpy.mockImplementationOnce(() => Promise.resolve(false));
       // @ts-expect-error ts fails to infer the return type because of a private method
       check1271SignatureBytesSpy.mockImplementationOnce(() => Promise.resolve(true));
 
-      // ethers.utils.formatBytes32String('approve rugpull')
+      // stringToHex('approve rugpull', { size: 32 })
       const message = sdkInstance.safe.calculateMessageHash(
         '0x617070726f76652072756770756c6c0000000000000000000000000000000000',
       );
@@ -333,22 +299,13 @@ describe('Safe Apps SDK safe methods', () => {
       const check1271SignatureSpy = jest.spyOn(sdkInstance.safe, 'check1271Signature');
       // @ts-expect-error method is private but we are testing it
       const check1271SignatureBytesSpy = jest.spyOn(sdkInstance.safe, 'check1271SignatureBytes');
-      safeInfoSpy.mockImplementationOnce(
-        (): Promise<SafeInfo> =>
-          Promise.resolve({
-            chainId: 4,
-            safeAddress: '0x9C6FEA0B2eAc5b6D8bBB6C30401D42aA95398190',
-            owners: [],
-            threshold: 1,
-            isReadOnly: false,
-          }),
-      );
+      safeInfoSpy.mockImplementationOnce((): Promise<SafeInfoExtended> => Promise.resolve(ExtendedSafeInfo));
       // @ts-expect-error ts fails to infer the return type because of a private method
       check1271SignatureSpy.mockImplementationOnce(() => Promise.resolve(false));
       // @ts-expect-error ts fails to infer the return type because of a private method
       check1271SignatureBytesSpy.mockImplementationOnce(() => Promise.resolve(false));
 
-      // ethers.utils.formatBytes32String('approve rugpull')
+      // stringToHex('approve rugpull', { size: 32 })
       const message = sdkInstance.safe.calculateMessageHash(
         '0x617070726f76652072756770756c6c0000000000000000000000000000000000',
       );
@@ -374,6 +331,17 @@ describe('Safe Apps SDK safe methods', () => {
       sdkInstance.safe.getChainInfo();
 
       expect(postMessageSpy).toHaveBeenCalledWith(expect.objectContaining({ method: Methods.getChainInfo }), '*');
+    });
+  });
+
+  describe('SDK.safe.getOffChainSignature', () => {
+    test('Should request the off-chain signature from the interface', () => {
+      sdkInstance.safe.getOffChainSignature('0x123');
+
+      expect(postMessageSpy).toHaveBeenCalledWith(
+        expect.objectContaining({ method: Methods.getOffChainSignature, params: '0x123' }),
+        '*',
+      );
     });
   });