+++ packages/cra-template-safe-app/CHANGELOG.md
@@ -1,4 +1,13 @@
-# @gnosis.pm/cra-template-safe-app
+# @safe-global/cra-template-safe-app
+
+## 5.0.0
+
+### Major Changes
+
+- 872b731: Upgrade CRA Safe App template
+  Use React 18
+  Use latest Safe Components theming and @mui/material v5
+  Remove `styled-components`
 
 ## 4.2.1